Rearranged lib paths - less annoying paths to declare. Pre-build libMad for Windows. Lots of size_t fixes.

+++ Fuji/Source/Drivers/D3D11/MFRenderer_D3D11.cpp
@@ -62,7 +62,7 @@ void MFRenderer_D3D11_SetDebugName(ID3D11DeviceChild* pResource, const char* pNa
 {
 #if !defined(MF_RETAIL)
 	if (pResource)
-		pResource->SetPrivateData(WKPDID_D3DDebugObjectName, MFString_Length(pName), pName);
+		pResource->SetPrivateData(WKPDID_D3DDebugObjectName, (UINT)MFString_Length(pName), pName);
 #endif
 }
