+++ Fuji/Source/Drivers/OpenGL/MFTexture_OpenGL.cpp
@@ -274,7 +274,7 @@ MF_API MFTexture* MFTexture_CreateRenderTarget(const char *pName, int width, int
 
 	if(!pTexture)
 	{
-		int nameLen = pName ? MFString_Length(pName) + 1 : 0;
+		size_t nameLen = pName ? MFString_Length(pName) + 1 : 0;
 		pTexture = (MFTexture*)MFHeap_Alloc(sizeof(MFTexture) + nameLen);
 
 		if(pName)
